Fix:add CRUD products operation

+++ src/products/products.controller.ts
@@ -2,10 +2,12 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Patch,
   Post,
+  Req,
   UseGuards,
 } from '@nestjs/common';
 import {
@@ -18,10 +20,10 @@ import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductResponseDto } from './dto/product-response.dto';
-import { Roles } from 'src/auth/decorators/roles.decorator';
-import { Role } from '@prisma/client';
+// import { Roles } from 'src/auth/decorators/roles.decorator';
+// import { Role } from '@prisma/client';
 import { AuthGuard } from '@nestjs/passport';
-import { RolesGuard } from 'src/auth/guards/roles.guard';
+// import { RolesGuard } from 'src/auth/guards/roles.guard';
 
 @ApiTags('products')
 @Controller('products')
@@ -52,8 +54,8 @@ export class ProductsController {
   }
 
   @Post()
-  @UseGuards(AuthGuard('jwt'), RolesGuard)
-  @Roles(Role.ADMIN)
+  @UseGuards(AuthGuard('jwt'))
+  // @Roles(Role.ADMIN)
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Create a new product' })
   @ApiResponse({
@@ -61,11 +63,22 @@ export class ProductsController {
     description: 'Product created successfully',
     type: ProductResponseDto,
   })
-  async createProduct(@Body() createProductDto: CreateProductDto) {
+  async createProduct(
+    @Body() createProductDto: CreateProductDto,
+    @Req() req: any,
+  ) {
+    //IMPLEMENT ROLE GUARD
+    const { role } = req.user;
+    if (role !== 'ADMIN') {
+      throw new ForbiddenException(
+        'Only admins are allowed to perform this action',
+      );
+    }
     return this.productsService.createProduct(createProductDto);
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard('jwt'))
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Update an existing product' })
   @ApiResponse({
@@ -77,16 +90,32 @@ export class ProductsController {
   async updateProduct(
     @Param('id') id: string,
     @Body() updateProductDto: UpdateProductDto,
+    @Req() req: any,
   ) {
+    //IMPLEMENT ROLE GUARD
+    const { role } = req['user'];
+    if (role !== 'ADMIN') {
+      throw new ForbiddenException(
+        'Only admins are allowed to perform this action',
+      );
+    }
     return this.productsService.updateProduct(id, updateProductDto);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard('jwt'))
   @ApiBearerAuth()
   @ApiOperation({ summary: 'Delete a product' })
   @ApiResponse({ status: 200, description: 'Product deleted successfully' })
   @ApiResponse({ status: 404, description: 'Product not found' })
-  async deleteProduct(@Param('id') id: string) {
+  async deleteProduct(@Param('id') id: string, @Req() req: any) {
+    //IMPLEMENT ROLE GUARD
+    const { role } = req['user'];
+    if (role !== 'ADMIN') {
+      throw new ForbiddenException(
+        'Only admins are allowed to perform this action',
+      );
+    }
     await this.productsService.deleteProduct(id);
     return { message: 'Product deleted successfully' };
   }
